+++ src/main/uni_hid_device.c
@@ -29,6 +29,7 @@ limitations under the License.
 #include "uni_hid_parser_android.h"
 #include "uni_hid_parser_ds3.h"
 #include "uni_hid_parser_ds4.h"
+#include "uni_hid_parser_ds5.h"
 #include "uni_hid_parser_generic.h"
 #include "uni_hid_parser_icade.h"
 #include "uni_hid_parser_nimbus.h"
@@ -443,6 +444,13 @@ void uni_hid_device_guess_controller_type_from_pid_vid(uni_hid_device_t* d) {
       d->report_parser.update_led = NULL;
 #endif  //
       logi("Device detected as DUALSHOCK4: 0x%02x\n", type);
+      break;
+    case CONTROLLER_TYPE_PS5Controller:
+      d->report_parser.setup = NULL;
+      d->report_parser.init_report = uni_hid_parser_ds5_init_report;
+      d->report_parser.parse_usage = uni_hid_parser_ds5_parse_usage;
+      d->report_parser.update_led = NULL;
+      logi("Device detected as DUALSHOCK5: 0x%02x\n", type);
       break;
     case CONTROLLER_TYPE_8BitdoController:
       d->report_parser.setup = NULL;